+++ core/src/fi/tuni/tiko/gameObject/tower/SportTower.java
@@ -21,7 +21,7 @@ public class SportTower extends Tower {
     private static final Sound sportSound = Gdx.audio.newSound(Gdx.files.internal("sounds/sport_sound.mp3"));
 
     public SportTower() {
-        super(new float[]{5, 6, 7}, new int[]{3, 4, 5}, new int[]{25, 45, 100});
+        super(new float[]{5, 6, 7}, new float[]{3, 4, 5}, new int[]{25, 45, 100});
     }
 
     public static SportTower getInstance() {
